module cleaning and rearanging
